Fix PHP8.1 fatal error unsupported operand types: string / int

+++ modules/Grades/Assignments.php
@@ -136,7 +136,8 @@ if ( ! empty( $_POST['tables'] ) )
 				elseif ( $column == 'FINAL_GRADE_PERCENT'
 					&& $table == 'GRADEBOOK_ASSIGNMENT_TYPES' )
 				{
-					$value = preg_replace( '/[^0-9.]/', '', $value ) / 100;
+					// Fix PHP8.1 fatal error unsupported operand types: string / int
+					$value = ( (float) preg_replace( '/[^0-9.]/', '', $value ) ) / 100;
 
 					if ( $value > 1 ) // 100%.
 					{
@@ -225,7 +226,8 @@ if ( ! empty( $_POST['tables'] ) )
 				elseif ( $column == 'FINAL_GRADE_PERCENT'
 					&& $table == 'GRADEBOOK_ASSIGNMENT_TYPES' )
 				{
-					$value = preg_replace( '/[^0-9.]/', '', $value ) / 100;
+					// Fix PHP8.1 fatal error unsupported operand types: string / int
+					$value = ( (float) preg_replace( '/[^0-9.]/', '', $value ) ) / 100;
 
 					if ( $value > 1 ) // 100%.
 					{
